fix(rust)!: return `u32` from `Node::child_count`
This matches the type returned from the underlying C library, and also
composes more cleanly with other functions like `Node::child`

+++ crates/cli/src/fuzz/corpus_test.rs
@@ -23,7 +23,7 @@ pub fn check_consistent_sizes(tree: &Tree, input: &[u8]) {
         let mut some_child_has_changes = false;
         let mut actual_named_child_count = 0;
         for i in 0..node.child_count() {
-            let child = node.child(i as u32).unwrap();
+            let child = node.child(i).unwrap();
             assert!(child.start_byte() >= last_child_end_byte);
             assert!(child.start_position() >= last_child_end_point);
             check(child, line_offsets);
